fix(ci): python-full-coverage — track post-handoff bail explicitly and split oversized test modules
The cooldown eligibility check treated any succeeded remonitor op as a post-finalize
start recheck bail, so persisted salvage cooldown never blocked reclaim after a real
monitor resume; split three test files that exceeded the 1500-line maintainability gate.
